Displaying the Favorite Meals in the app by using the context API. Also refactored the list of meals into a individual component to reuse it in the application.
