+++ screens/MealsOverviewScreen.js
@@ -1,8 +1,8 @@
-import { View, StyleSheet, FlatList } from "react-native";
 import { useNavigation, useRoute } from "@react-navigation/native";
 import { CATEGORIES, MEALS } from "../data/dummy_data";
-import IndividualMealItem from "../components/IndividualMealItem";
 import { useLayoutEffect } from "react";
+import MealList from "../components/MealList/MealList";
+
 
 function MealsOverviewScreen(){
 
@@ -20,35 +20,8 @@ function MealsOverviewScreen(){
         })
     }, [catId, navigation]);
 
-    function displayMealItem(itemData){
-        const item = itemData.item;
-        const mealItemProps = {
-            id: item.id,
-            title: item.title,
-            imageUrl: item.imageUrl,
-            duration: item.duration,
-            complexity: item.complexity,
-            affordability: item.affordability,
-        }
-        return <IndividualMealItem {...mealItemProps} onPressMealCard={()=>navigation.navigate('MealDetails', {mealId: mealItemProps.id})}/>
-    }
-    
-    return (
-        <View style={styles.rootContainer}>
-            <FlatList
-                data={displayedMeals} 
-                keyExtractor={(item) => item.id}
-                renderItem={displayMealItem}
-                />
-        </View>
-    );
+    return <MealList items={displayedMeals}/>
 }
 
 export default MealsOverviewScreen;
 
-const styles = StyleSheet.create({
-    rootContainer: {
-        flex: 1,
-        padding: 16,
-    }
-})
